Fixed for Ia64 + delayed key creation + a lot of small bug fixes
Docs/Makefile.am:
  Removed typo
Docs/internals.texi:
  Added documentation about filesort
Docs/manual.texi:
  Lots of changes, including information about the password in the mysql.user table
config.h.in:
  ***MISSING WEAVE***
heap/hp_rnext.c:
  Fix of reading of read-next
include/config-win.h:
  Fix for Ia64
include/global.h:
  Fix for Ia64
include/my_base.h:
  Fix for Ia64
include/myisam.h:
  New repair options
isam/_dynrec.c:
  Typo fixes
isam/_search.c:
  Fix for Ia64
merge/rrnd.c:
  Fix of ORDER BY for MERGE tables
myisam/ft_eval.c:
  Fix for Ia64
myisam/ft_parser.c:
  Fix for Ia64
myisam/ft_static.c:
  Portability fixes
myisam/mi_check.c:
  Fix for Ia64 + delay key creation
myisam/mi_locking.c:
  Portability fixes
myisam/myisamdef.h:
  Fixes for automatic table repair
myisam/myisampack.c:
  Version number increment
myisammrg/myrg_rrnd.c:
  Fix of ORDER BY for mymerge tables
mysql.proj:
  Updated
mysys/charset.c:
  Fix for Ia64
mysys/mf_casecnv.c:
  Fix for Ia64
mysys/mf_pack.c:
  Fix for Ia64
mysys/mf_path.c:
  Fix for Ia64
mysys/mf_wfile.c:
  Fix for Ia64
mysys/my_alloc.c:
  Fix for Ia64
mysys/my_lib.c:
  Fix for Ia64
mysys/my_lread.c:
  Fix for Ia64
mysys/my_lwrite.c:
  Fix for Ia64
mysys/my_winthread.c:
  Fix for Ia64
mysys/string.c:
  Fix for Ia64
mysys/thr_alarm.c:
  Fix for Ia64
mysys/thr_rwlock.c:
  Fix of scheduler bug for systems without rw_locks
regex/regcomp.c:
  Fix for Ia64
regex/regex.h:
  Fix for Ia64
regex/regex2.h:
  Fix for Ia64
scripts/mysqlhotcopy.sh:
  Explicit version number (because of bitkeeper)
sql-bench/bench-init.pl.sh:
  New help message
sql-bench/server-cfg.sh:
  Updates for postgreSQL 7.0
sql/ChangeLog:
  Changelog
sql/field.cc:
  Fix of datetime function handling
sql/filesort.cc:
  More comments
sql/gen_lex_hash.cc:
  More options
sql/ha_berkeley.cc:
  Fix for part key reading
sql/ha_berkeley.h:
  Indentation
sql/ha_myisam.cc:
  Fix for delayed key creation
sql/ha_myisam.h:
  Fix for delayed key creation
sql/handler.h:
  Fix for delayed key creation
sql/item_func.cc:
  Fix of crash when using constant expression with UDF functions in WHERE
sql/item_timefunc.cc:
  Fix of datetime function handling
sql/mini_client.cc:
  Portability fix
sql/mysql_priv.h:
  Optimization of select initialization
sql/mysqld.cc:
  Portability fixes, thread cache and min argument for connect_timeout
sql/sql_load.cc:
  Delayed key handling
sql/sql_parse.cc:
  Portability fixes and select initialization optimization
sql/sql_select.cc:
  Small optimization + more comments
sql/sql_table.cc:
  Fix default value adding to blob with alter table + delayed index creation
sql/sql_yacc.yy:
  Optimized select initialization
sql/structs.h:
  Fix for part key read
strings/ctype-sjis.c:
  Fix for Ia64
strings/ctype-tis620.c:
  Fix for Ia64
strings/t_ctype.h:
  bk fix
support-files/mysql.spec.sh:
  Added comment
vio/viotest-ssl.cc:
  bk fix
vio/viotest-sslconnect.cc:
  bk fix
vio/viotest.cc:
  bk fix
BitKeeper/etc/logging_ok:
  Logging to logging@openlogging.org accepted

+++ isam/_search.c
@@ -342,7 +342,7 @@ int _nisam_key_cmp(register N_KEYSEG *keyseg, register uchar *a, register uchar 
           else
           {
             if ((flag = my_strnncoll(default_charset_info,
-				     a, (end-a), b, b_length)))
+				     a, (int) (end-a), b, b_length)))
               return (keyseg->base.flag & HA_REVERSE_SORT) ? -flag : flag;
             b+= (uint) (end-a);
             a=end;
@@ -393,7 +393,7 @@ int _nisam_key_cmp(register N_KEYSEG *keyseg, register uchar *a, register uchar 
           else
           {
             if ((flag = my_strnncoll(default_charset_info,
-				     a, (end-a), b, (end-a))))
+				     a, (int) (end-a), b, (int) (end-a))))
               return (keyseg->base.flag & HA_REVERSE_SORT) ? -flag : flag;
             b+= (uint) (end-a);
             a=end;
